Posse-post-discovery: use rel=feed and type=text/html
When looking for the author's h-feed url, make sure to
check the content-type. If it is not given, uses
follow_redirects to resolve the url and get the content-type
from the response headers.

Added unit tests for cases where rel=feed is not of
type text/html. if the type is given, we should not follow
non-text/html feeds. if the type is not given, we fetch each feed
until we find one that matches the expected type.

Added a few lines to util.follow_redirects reraise
AssertionErrors to facilitate better testing.