original post discovery: store rel-syndication links in upstreamDuplicates, not tags
part of snarfed/bridgy#154. we're going to start considering PSCs and rel-syndication links "explicit" original post links, as opposed to links in attachments and text, which are weaker signals.

next step is to handle upstreamDuplicates in propagate and include them as webmention targets.

cc @kylewm. mind sanity checking this for me?