store last succesful get_activities_response() in every source
needed so we can use it to determine when an activity has new comments/likes/etc without depending on memcache (and thus failing when it evicts). for #22; details in https://github.com/snarfed/bridgy/issues/22#issuecomment-56279420

!!! NOT ENTIRELY SAFE !!! this sometimes raises `ValueError: Circular reference detected` on real world G+ data, probably due to how i hack up the inner objects during a poll. whee.

don't worry, this will get reverted immediately and never deployed. i'm only pushing it to origin for posterity.