start on support for facebook pages. for #41
this detects whether the user who's signing up is an admin of any pages, and if they are, shows a page for them to choose whether to sign up their own account or a page. based on the similar existing page for choosing btw blogger/tumblr blogs.

adds the manage_page scope to the facebook oauth dance, which includes write permissions. sadly there's no read only permission to get a user's pages. :(

https://developers.facebook.com/docs/graph-api/reference/v2.2/user/accounts