cache webmention endpoint discovery errors; drop cache expiration to 2h
before this, we only cached webmention endpoint discovery results when we successfully fetched the page. this is a bit aggressive, since it means we'll cache transient network failures, but with the lower expiration (2h instead of 1d), and now that the retry button flushes the cache (#524), i think it's ok.

for #534