testutil refactoring: merge FakeBase into FakeSource, drop it from FakeGrSource
this fixes the tests when we disable the ndb instance for #558. the culprit was the INDEPENDENT datastore transaction that we used to store a FakeGrSource in FakeSource's ctor.

FakeGrSource needed FakeBase (and therefore to be a datastore entity) when FakeSource overrode all the get_*() methods, but it doesn't now that i've dropped those in #404. than god.

i'd still like to get rid of string_id_counter at some point, but this is still a nice simplification. the old code was way overengineered.