Fix: Request Publish scopes on sign-in
If we don't request `PUBLISH_SCOPES` on the initial login, you then have
to go through the Publish authorization flow before RSVPing to an event,
otherwise you'll receive 400s with `insufficient_scope`.

This isn't the nicest UX, so we should instead request the
`PUBLISH_SCOPES` up-front, as we don't support Listen, anyway.

Follow-up to #873.