Fix: Escape HTML tags in content
As part of #880, we added a check to ensure that we would not be subject
to Cross-Site Scripting attacks (XSS).

However, as noted in #936, a post that does contain HTML is still being
rejected.

The simplest way to get around this is to escape the tags with the HTML
entities corresponding with them, so they can be safely rendered.

Closes #936.