fix propagate bug that makes us lose targets during failures like #237
discovered while debugging #1124. subtle one! we modify the entity in memory as we iterate through the targets and attempt each one, and we handle most exceptions and continue on to the next target, but not all. during #237 and related errors, we short circuit out, and notably, we lose the target we're currently on. this fixes that.