always populate Response.activities_json, regardless of response type etc
fixes #237...but also this is dumb, brute force. I'm not at all confident in it, I'm sure the previous logic for only sometimes populating activities_json was here for a reason, but I'm not sure why, and all the tests pass, and I think the only resulting behavior change (other than fixing #237) would be that we're now maybe even more promiscuous in discovering webmention targets. eg I think for normal replies, this may make us try to send wms to URLs in the reply text, which we maybe didn't do before. that should be harmless.