first major fix for empty Response.activities_json (ie KeyError bug)
for #237. when determining the source post's URL in PropagateTask.source_url(), if Response.activities_json is empty, this uses the response object from response_json.

this keeps us from crashing in propagate tasks, but it doesn't fully fix the bug. with this, reply tweets that link to the user's site now miss the link as a webmention target, when it should be included. still need to come up with a test that reproduces that.